pyproject.toml
[project]
name = "ultrasound-imaging-toolkit"
version = "0.1.0"
description = "Ultrasound Image Processing and Analysis Toolkit - Demonstrating signal processing, image analysis, and ML for medical ultrasound applications"
authors = [
    {name = "Reza Mirzaeifard", email = "reza.mirzaeifard@gmail.com"}
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.3.0",
    "opencv-python-headless>=4.8.0",
    "torch>=2.0.0,<2.3.0",
    "torchvision>=0.15.0,<0.18.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "kaggle>=1.5.0",
    "jupyter>=1.0.0",
    "seaborn>=0.12.0",
    "ipykernel>=6.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/rezamirzaei/ultrasound-imaging-toolkit"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.uv]
# Ensure uv resolves to versions with compatible wheels for all platforms
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux'",
    "sys_platform == 'win32'",
]
